Service and coverage level metadata

Get metadata objects from a WCSClient object. <WCSCoverageSummary> in particular can be used to extract further metadata about individual coverages.

Usage

emdn_get_coverage_summaries(wcs, coverage_ids)

emdn_get_coverage_summaries_all(wcs)

emdn_get_coverage_ids(wcs)

emdn_has_dimension(wcs, coverage_ids, type = c("temporal", "vertical"))

emdn_get_coverage_dim_coefs(
  wcs,
  coverage_ids,
  type = c("temporal", "vertical")
)

Arguments

wcs

A WCSClient R6 object, created with function emdn_init_wcs_client.

coverage_ids

character vector of coverage IDs.

type

a coverage dimension type. One of "temporal" or "vertical".

Value

emdn_get_coverage_summaries: returns a list of objects of class <WCSCoverageSummary> for each coverage_id provided.

emdn_get_coverage_summaries_all: returns a list of objects of class <WCSCoverageSummary> for each coverage avalable through the service.

emdn_get_coverage_ids returns a character vector of coverage ids.

emdn_get_coverage_dim_coefs returns a list containing a vector of coefficients for each coverage requested.

a list containing a vector of coefficients for each coverage requested.

Functions

emdn_get_coverage_summaries(): Get summaries for specific coverages.

emdn_get_coverage_summaries_all(): Get summaries for all available coverages from a service.

emdn_get_coverage_ids(): Get coverage IDs for all available coverages from a service.

emdn_has_dimension(): check whether a coverage has a temporal or vertical dimension.

emdn_get_coverage_dim_coefs(): Get temporal or vertical coefficients for a coverage.

Examples

wcs <- emdn_init_wcs_client(service = "biology")
#> ✔ WCS client created succesfully
#> ℹ Service: <https://geo.vliz.be/geoserver/Emodnetbio/wcs>
#> ℹ Service: "2.0.1"
cov_ids <- emdn_get_coverage_ids(wcs)
cov_ids
#>  [1] "Emodnetbio__ratio_large_to_small_19582016_L1_err"
#>  [2] "Emodnetbio__aca_spp_19582016_L1"                 
#>  [3] "Emodnetbio__cal_fin_19582016_L1_err"             
#>  [4] "Emodnetbio__cal_hel_19582016_L1_err"             
#>  [5] "Emodnetbio__met_luc_19582016_L1_err"             
#>  [6] "Emodnetbio__oit_spp_19582016_L1_err"             
#>  [7] "Emodnetbio__tem_lon_19582016_L1_err"             
#>  [8] "Emodnetbio__chli_19582016_L1_err"                
#>  [9] "Emodnetbio__tot_lar_19582016_L1_err"             
#> [10] "Emodnetbio__tot_sma_19582016_L1_err"             
emdn_has_dimension(wcs,
  cov_ids,
  type = "temporal"
)
#> No encoding supplied: defaulting to UTF-8.
#> No encoding supplied: defaulting to UTF-8.
#> No encoding supplied: defaulting to UTF-8.
#> No encoding supplied: defaulting to UTF-8.
#> No encoding supplied: defaulting to UTF-8.
#> No encoding supplied: defaulting to UTF-8.
#> No encoding supplied: defaulting to UTF-8.
#> No encoding supplied: defaulting to UTF-8.
#> No encoding supplied: defaulting to UTF-8.
#> No encoding supplied: defaulting to UTF-8.
#> Emodnetbio__ratio_large_to_small_19582016_L1_err 
#>                                             TRUE 
#>                  Emodnetbio__aca_spp_19582016_L1 
#>                                             TRUE 
#>              Emodnetbio__cal_fin_19582016_L1_err 
#>                                             TRUE 
#>              Emodnetbio__cal_hel_19582016_L1_err 
#>                                             TRUE 
#>              Emodnetbio__met_luc_19582016_L1_err 
#>                                             TRUE 
#>              Emodnetbio__oit_spp_19582016_L1_err 
#>                                             TRUE 
#>              Emodnetbio__tem_lon_19582016_L1_err 
#>                                             TRUE 
#>                 Emodnetbio__chli_19582016_L1_err 
#>                                             TRUE 
#>              Emodnetbio__tot_lar_19582016_L1_err 
#>                                             TRUE 
#>              Emodnetbio__tot_sma_19582016_L1_err 
#>                                             TRUE 
emdn_has_dimension(wcs,
  cov_ids,
  type = "vertical"
)
#> Emodnetbio__ratio_large_to_small_19582016_L1_err 
#>                                            FALSE 
#>                  Emodnetbio__aca_spp_19582016_L1 
#>                                            FALSE 
#>              Emodnetbio__cal_fin_19582016_L1_err 
#>                                            FALSE 
#>              Emodnetbio__cal_hel_19582016_L1_err 
#>                                            FALSE 
#>              Emodnetbio__met_luc_19582016_L1_err 
#>                                            FALSE 
#>              Emodnetbio__oit_spp_19582016_L1_err 
#>                                            FALSE 
#>              Emodnetbio__tem_lon_19582016_L1_err 
#>                                            FALSE 
#>                 Emodnetbio__chli_19582016_L1_err 
#>                                            FALSE 
#>              Emodnetbio__tot_lar_19582016_L1_err 
#>                                            FALSE 
#>              Emodnetbio__tot_sma_19582016_L1_err 
#>                                            FALSE 
emdn_get_coverage_summaries(wcs, cov_ids[1:2])
#> [[1]]
#> <WCSCoverageSummary>
#> ....|-- CoverageId: Emodnetbio__ratio_large_to_small_19582016_L1_err
#> ....|-- CoverageSubtype: RectifiedGridCoverage
#> ....|-- WGS84BoundingBox <OWSWGS84BoundingBox>
#> ........|-- LowerCorner: -75.05 34.95
#> ........|-- UpperCorner: 20.05 75.05
#> ....|-- BoundingBox  [crs=http://www.opengis.net/def/crs/EPSG/0/4326] <OWSBoundingBox>
#> ........|-- LowerCorner: -75.05 34.95
#> ........|-- UpperCorner: 20.05 75.05
#> [[2]]
#> <WCSCoverageSummary>
#> ....|-- CoverageId: Emodnetbio__aca_spp_19582016_L1
#> ....|-- CoverageSubtype: RectifiedGridCoverage
#> ....|-- WGS84BoundingBox <OWSWGS84BoundingBox>
#> ........|-- LowerCorner: -75.05 34.95
#> ........|-- UpperCorner: 20.05 75.05
#> ....|-- BoundingBox  [crs=http://www.opengis.net/def/crs/EPSG/0/4326] <OWSBoundingBox>
#> ........|-- LowerCorner: -75.05 34.95
#> ........|-- UpperCorner: 20.05 75.05
emdn_get_coverage_summaries_all(wcs)
#> [[1]]
#> <WCSCoverageSummary>
#> ....|-- CoverageId: Emodnetbio__ratio_large_to_small_19582016_L1_err
#> ....|-- CoverageSubtype: RectifiedGridCoverage
#> ....|-- WGS84BoundingBox <OWSWGS84BoundingBox>
#> ........|-- LowerCorner: -75.05 34.95
#> ........|-- UpperCorner: 20.05 75.05
#> ....|-- BoundingBox  [crs=http://www.opengis.net/def/crs/EPSG/0/4326] <OWSBoundingBox>
#> ........|-- LowerCorner: -75.05 34.95
#> ........|-- UpperCorner: 20.05 75.05
#> [[2]]
#> <WCSCoverageSummary>
#> ....|-- CoverageId: Emodnetbio__aca_spp_19582016_L1
#> ....|-- CoverageSubtype: RectifiedGridCoverage
#> ....|-- WGS84BoundingBox <OWSWGS84BoundingBox>
#> ........|-- LowerCorner: -75.05 34.95
#> ........|-- UpperCorner: 20.05 75.05
#> ....|-- BoundingBox  [crs=http://www.opengis.net/def/crs/EPSG/0/4326] <OWSBoundingBox>
#> ........|-- LowerCorner: -75.05 34.95
#> ........|-- UpperCorner: 20.05 75.05
#> [[3]]
#> <WCSCoverageSummary>
#> ....|-- CoverageId: Emodnetbio__cal_fin_19582016_L1_err
#> ....|-- CoverageSubtype: RectifiedGridCoverage
#> ....|-- WGS84BoundingBox <OWSWGS84BoundingBox>
#> ........|-- LowerCorner: -75.05 34.95
#> ........|-- UpperCorner: 20.05 75.05
#> ....|-- BoundingBox  [crs=http://www.opengis.net/def/crs/EPSG/0/4326] <OWSBoundingBox>
#> ........|-- LowerCorner: -75.05 34.95
#> ........|-- UpperCorner: 20.05 75.05
#> [[4]]
#> <WCSCoverageSummary>
#> ....|-- CoverageId: Emodnetbio__cal_hel_19582016_L1_err
#> ....|-- CoverageSubtype: RectifiedGridCoverage
#> ....|-- WGS84BoundingBox <OWSWGS84BoundingBox>
#> ........|-- LowerCorner: -75.05 34.95
#> ........|-- UpperCorner: 20.05 75.05
#> ....|-- BoundingBox  [crs=http://www.opengis.net/def/crs/EPSG/0/4326] <OWSBoundingBox>
#> ........|-- LowerCorner: -75.05 34.95
#> ........|-- UpperCorner: 20.05 75.05
#> [[5]]
#> <WCSCoverageSummary>
#> ....|-- CoverageId: Emodnetbio__met_luc_19582016_L1_err
#> ....|-- CoverageSubtype: RectifiedGridCoverage
#> ....|-- WGS84BoundingBox <OWSWGS84BoundingBox>
#> ........|-- LowerCorner: -75.05 34.95
#> ........|-- UpperCorner: 20.05 75.05
#> ....|-- BoundingBox  [crs=http://www.opengis.net/def/crs/EPSG/0/4326] <OWSBoundingBox>
#> ........|-- LowerCorner: -75.05 34.95
#> ........|-- UpperCorner: 20.05 75.05
#> [[6]]
#> <WCSCoverageSummary>
#> ....|-- CoverageId: Emodnetbio__oit_spp_19582016_L1_err
#> ....|-- CoverageSubtype: RectifiedGridCoverage
#> ....|-- WGS84BoundingBox <OWSWGS84BoundingBox>
#> ........|-- LowerCorner: -75.05 34.95
#> ........|-- UpperCorner: 20.05 75.05
#> ....|-- BoundingBox  [crs=http://www.opengis.net/def/crs/EPSG/0/4326] <OWSBoundingBox>
#> ........|-- LowerCorner: -75.05 34.95
#> ........|-- UpperCorner: 20.05 75.05
#> [[7]]
#> <WCSCoverageSummary>
#> ....|-- CoverageId: Emodnetbio__tem_lon_19582016_L1_err
#> ....|-- CoverageSubtype: RectifiedGridCoverage
#> ....|-- WGS84BoundingBox <OWSWGS84BoundingBox>
#> ........|-- LowerCorner: -75.05 34.95
#> ........|-- UpperCorner: 20.05 75.05
#> ....|-- BoundingBox  [crs=http://www.opengis.net/def/crs/EPSG/0/4326] <OWSBoundingBox>
#> ........|-- LowerCorner: -75.05 34.95
#> ........|-- UpperCorner: 20.05 75.05
#> [[8]]
#> <WCSCoverageSummary>
#> ....|-- CoverageId: Emodnetbio__chli_19582016_L1_err
#> ....|-- CoverageSubtype: RectifiedGridCoverage
#> ....|-- WGS84BoundingBox <OWSWGS84BoundingBox>
#> ........|-- LowerCorner: -75.05 34.95
#> ........|-- UpperCorner: 20.05 75.05
#> ....|-- BoundingBox  [crs=http://www.opengis.net/def/crs/EPSG/0/4326] <OWSBoundingBox>
#> ........|-- LowerCorner: -75.05 34.95
#> ........|-- UpperCorner: 20.05 75.05
#> [[9]]
#> <WCSCoverageSummary>
#> ....|-- CoverageId: Emodnetbio__tot_lar_19582016_L1_err
#> ....|-- CoverageSubtype: RectifiedGridCoverage
#> ....|-- WGS84BoundingBox <OWSWGS84BoundingBox>
#> ........|-- LowerCorner: -75.05 34.95
#> ........|-- UpperCorner: 20.05 75.05
#> ....|-- BoundingBox  [crs=http://www.opengis.net/def/crs/EPSG/0/4326] <OWSBoundingBox>
#> ........|-- LowerCorner: -75.05 34.95
#> ........|-- UpperCorner: 20.05 75.05
#> [[10]]
#> <WCSCoverageSummary>
#> ....|-- CoverageId: Emodnetbio__tot_sma_19582016_L1_err
#> ....|-- CoverageSubtype: RectifiedGridCoverage
#> ....|-- WGS84BoundingBox <OWSWGS84BoundingBox>
#> ........|-- LowerCorner: -75.05 34.95
#> ........|-- UpperCorner: 20.05 75.05
#> ....|-- BoundingBox  [crs=http://www.opengis.net/def/crs/EPSG/0/4326] <OWSBoundingBox>
#> ........|-- LowerCorner: -75.05 34.95
#> ........|-- UpperCorner: 20.05 75.05
emdn_get_coverage_dim_coefs(wcs,
  cov_ids[1:2],
  type = "temporal"
)
#> $Emodnetbio__ratio_large_to_small_19582016_L1_err
#>   [1] "1958-02-16T01:00:00" "1958-05-16T01:00:00" "1958-08-16T01:00:00"
#>   [4] "1958-11-16T01:00:00" "1959-02-16T01:00:00" "1959-05-16T01:00:00"
#>   [7] "1959-08-16T01:00:00" "1959-11-16T01:00:00" "1960-02-16T01:00:00"
#>  [10] "1960-05-16T01:00:00" "1960-08-16T01:00:00" "1960-11-16T01:00:00"
#>  [13] "1961-02-16T01:00:00" "1961-05-16T01:00:00" "1961-08-16T01:00:00"
#>  [16] "1961-11-16T01:00:00" "1962-02-16T01:00:00" "1962-05-16T01:00:00"
#>  [19] "1962-08-16T01:00:00" "1962-11-16T01:00:00" "1963-02-16T01:00:00"
#>  [22] "1963-05-16T01:00:00" "1963-08-16T01:00:00" "1963-11-16T01:00:00"
#>  [25] "1964-02-16T01:00:00" "1964-05-16T01:00:00" "1964-08-16T01:00:00"
#>  [28] "1964-11-16T01:00:00" "1965-02-16T01:00:00" "1965-05-16T01:00:00"
#>  [31] "1965-08-16T01:00:00" "1965-11-16T01:00:00" "1966-02-16T01:00:00"
#>  [34] "1966-05-16T01:00:00" "1966-08-16T01:00:00" "1966-11-16T01:00:00"
#>  [37] "1967-02-16T01:00:00" "1967-05-16T01:00:00" "1967-08-16T01:00:00"
#>  [40] "1967-11-16T01:00:00" "1968-02-16T01:00:00" "1968-05-16T01:00:00"
#>  [43] "1968-08-16T01:00:00" "1968-11-16T01:00:00" "1969-02-16T01:00:00"
#>  [46] "1969-05-16T01:00:00" "1969-08-16T01:00:00" "1969-11-16T01:00:00"
#>  [49] "1970-02-16T01:00:00" "1970-05-16T01:00:00" "1970-08-16T01:00:00"
#>  [52] "1970-11-16T01:00:00" "1971-02-16T01:00:00" "1971-05-16T01:00:00"
#>  [55] "1971-08-16T01:00:00" "1971-11-16T01:00:00" "1972-02-16T01:00:00"
#>  [58] "1972-05-16T01:00:00" "1972-08-16T01:00:00" "1972-11-16T01:00:00"
#>  [61] "1973-02-16T01:00:00" "1973-05-16T01:00:00" "1973-08-16T01:00:00"
#>  [64] "1973-11-16T01:00:00" "1974-02-16T01:00:00" "1974-05-16T01:00:00"
#>  [67] "1974-08-16T01:00:00" "1974-11-16T01:00:00" "1975-02-16T01:00:00"
#>  [70] "1975-05-16T01:00:00" "1975-08-16T01:00:00" "1975-11-16T01:00:00"
#>  [73] "1976-02-16T01:00:00" "1976-05-16T01:00:00" "1976-08-16T01:00:00"
#>  [76] "1976-11-16T01:00:00" "1977-02-16T01:00:00" "1977-05-16T02:00:00"
#>  [79] "1977-08-16T02:00:00" "1977-11-16T01:00:00" "1978-02-16T01:00:00"
#>  [82] "1978-05-16T02:00:00" "1978-08-16T02:00:00" "1978-11-16T01:00:00"
#>  [85] "1979-02-16T01:00:00" "1979-05-16T02:00:00" "1979-08-16T02:00:00"
#>  [88] "1979-11-16T01:00:00" "1980-02-16T01:00:00" "1980-05-16T02:00:00"
#>  [91] "1980-08-16T02:00:00" "1980-11-16T01:00:00" "1981-02-16T01:00:00"
#>  [94] "1981-05-16T02:00:00" "1981-08-16T02:00:00" "1981-11-16T01:00:00"
#>  [97] "1982-02-16T01:00:00" "1982-05-16T02:00:00" "1982-08-16T02:00:00"
#> [100] "1982-11-16T01:00:00" "1983-02-16T01:00:00" "1983-05-16T02:00:00"
#> [103] "1983-08-16T02:00:00" "1983-11-16T01:00:00" "1984-02-16T01:00:00"
#> [106] "1984-05-16T02:00:00" "1984-08-16T02:00:00" "1984-11-16T01:00:00"
#> [109] "1985-02-16T01:00:00" "1985-05-16T02:00:00" "1985-08-16T02:00:00"
#> [112] "1985-11-16T01:00:00" "1986-02-16T01:00:00" "1986-05-16T02:00:00"
#> [115] "1986-08-16T02:00:00" "1986-11-16T01:00:00" "1987-02-16T01:00:00"
#> [118] "1987-05-16T02:00:00" "1987-08-16T02:00:00" "1987-11-16T01:00:00"
#> [121] "1988-02-16T01:00:00" "1988-05-16T02:00:00" "1988-08-16T02:00:00"
#> [124] "1988-11-16T01:00:00" "1989-02-16T01:00:00" "1989-05-16T02:00:00"
#> [127] "1989-08-16T02:00:00" "1989-11-16T01:00:00" "1990-02-16T01:00:00"
#> [130] "1990-05-16T02:00:00" "1990-08-16T02:00:00" "1990-11-16T01:00:00"
#> [133] "1991-02-16T01:00:00" "1991-05-16T02:00:00" "1991-08-16T02:00:00"
#> [136] "1991-11-16T01:00:00" "1992-02-16T01:00:00" "1992-05-16T02:00:00"
#> [139] "1992-08-16T02:00:00" "1992-11-16T01:00:00" "1993-02-16T01:00:00"
#> [142] "1993-05-16T02:00:00" "1993-08-16T02:00:00" "1993-11-16T01:00:00"
#> [145] "1994-02-16T01:00:00" "1994-05-16T02:00:00" "1994-08-16T02:00:00"
#> [148] "1994-11-16T01:00:00" "1995-02-16T01:00:00" "1995-05-16T02:00:00"
#> [151] "1995-08-16T02:00:00" "1995-11-16T01:00:00" "1996-02-16T01:00:00"
#> [154] "1996-05-16T02:00:00" "1996-08-16T02:00:00" "1996-11-16T01:00:00"
#> [157] "1997-02-16T01:00:00" "1997-05-16T02:00:00" "1997-08-16T02:00:00"
#> [160] "1997-11-16T01:00:00" "1998-02-16T01:00:00" "1998-05-16T02:00:00"
#> [163] "1998-08-16T02:00:00" "1998-11-16T01:00:00" "1999-02-16T01:00:00"
#> [166] "1999-05-16T02:00:00" "1999-08-16T02:00:00" "1999-11-16T01:00:00"
#> [169] "2000-02-16T01:00:00" "2000-05-16T02:00:00" "2000-08-16T02:00:00"
#> [172] "2000-11-16T01:00:00" "2001-02-16T01:00:00" "2001-05-16T02:00:00"
#> [175] "2001-08-16T02:00:00" "2001-11-16T01:00:00" "2002-02-16T01:00:00"
#> [178] "2002-05-16T02:00:00" "2002-08-16T02:00:00" "2002-11-16T01:00:00"
#> [181] "2003-02-16T01:00:00" "2003-05-16T02:00:00" "2003-08-16T02:00:00"
#> [184] "2003-11-16T01:00:00" "2004-02-16T01:00:00" "2004-05-16T02:00:00"
#> [187] "2004-08-16T02:00:00" "2004-11-16T01:00:00" "2005-02-16T01:00:00"
#> [190] "2005-05-16T02:00:00" "2005-08-16T02:00:00" "2005-11-16T01:00:00"
#> [193] "2006-02-16T01:00:00" "2006-05-16T02:00:00" "2006-08-16T02:00:00"
#> [196] "2006-11-16T01:00:00" "2007-02-16T01:00:00" "2007-05-16T02:00:00"
#> [199] "2007-08-16T02:00:00" "2007-11-16T01:00:00" "2008-02-16T01:00:00"
#> [202] "2008-05-16T02:00:00" "2008-08-16T02:00:00" "2008-11-16T01:00:00"
#> [205] "2009-02-16T01:00:00" "2009-05-16T02:00:00" "2009-08-16T02:00:00"
#> [208] "2009-11-16T01:00:00" "2010-02-16T01:00:00" "2010-05-16T02:00:00"
#> [211] "2010-08-16T02:00:00" "2010-11-16T01:00:00" "2011-02-16T01:00:00"
#> [214] "2011-05-16T02:00:00" "2011-08-16T02:00:00" "2011-11-16T01:00:00"
#> [217] "2012-02-16T01:00:00" "2012-05-16T02:00:00" "2012-08-16T02:00:00"
#> [220] "2012-11-16T01:00:00" "2013-02-16T01:00:00" "2013-05-16T02:00:00"
#> [223] "2013-08-16T02:00:00" "2013-11-16T01:00:00" "2014-02-16T01:00:00"
#> [226] "2014-05-16T02:00:00" "2014-08-16T02:00:00" "2014-11-16T01:00:00"
#> [229] "2015-02-16T01:00:00" "2015-05-16T02:00:00" "2015-08-16T02:00:00"
#> [232] "2015-11-16T01:00:00" "2016-02-16T01:00:00" "2016-05-16T02:00:00"
#> [235] "2016-08-16T02:00:00" "2016-11-16T01:00:00"
#> attr(,"type")
#> temporal_coefficents
#> 
#> $Emodnetbio__aca_spp_19582016_L1
#>   [1] "1958-02-16T01:00:00" "1958-05-16T01:00:00" "1958-08-16T01:00:00"
#>   [4] "1958-11-16T01:00:00" "1959-02-16T01:00:00" "1959-05-16T01:00:00"
#>   [7] "1959-08-16T01:00:00" "1959-11-16T01:00:00" "1960-02-16T01:00:00"
#>  [10] "1960-05-16T01:00:00" "1960-08-16T01:00:00" "1960-11-16T01:00:00"
#>  [13] "1961-02-16T01:00:00" "1961-05-16T01:00:00" "1961-08-16T01:00:00"
#>  [16] "1961-11-16T01:00:00" "1962-02-16T01:00:00" "1962-05-16T01:00:00"
#>  [19] "1962-08-16T01:00:00" "1962-11-16T01:00:00" "1963-02-16T01:00:00"
#>  [22] "1963-05-16T01:00:00" "1963-08-16T01:00:00" "1963-11-16T01:00:00"
#>  [25] "1964-02-16T01:00:00" "1964-05-16T01:00:00" "1964-08-16T01:00:00"
#>  [28] "1964-11-16T01:00:00" "1965-02-16T01:00:00" "1965-05-16T01:00:00"
#>  [31] "1965-08-16T01:00:00" "1965-11-16T01:00:00" "1966-02-16T01:00:00"
#>  [34] "1966-05-16T01:00:00" "1966-08-16T01:00:00" "1966-11-16T01:00:00"
#>  [37] "1967-02-16T01:00:00" "1967-05-16T01:00:00" "1967-08-16T01:00:00"
#>  [40] "1967-11-16T01:00:00" "1968-02-16T01:00:00" "1968-05-16T01:00:00"
#>  [43] "1968-08-16T01:00:00" "1968-11-16T01:00:00" "1969-02-16T01:00:00"
#>  [46] "1969-05-16T01:00:00" "1969-08-16T01:00:00" "1969-11-16T01:00:00"
#>  [49] "1970-02-16T01:00:00" "1970-05-16T01:00:00" "1970-08-16T01:00:00"
#>  [52] "1970-11-16T01:00:00" "1971-02-16T01:00:00" "1971-05-16T01:00:00"
#>  [55] "1971-08-16T01:00:00" "1971-11-16T01:00:00" "1972-02-16T01:00:00"
#>  [58] "1972-05-16T01:00:00" "1972-08-16T01:00:00" "1972-11-16T01:00:00"
#>  [61] "1973-02-16T01:00:00" "1973-05-16T01:00:00" "1973-08-16T01:00:00"
#>  [64] "1973-11-16T01:00:00" "1974-02-16T01:00:00" "1974-05-16T01:00:00"
#>  [67] "1974-08-16T01:00:00" "1974-11-16T01:00:00" "1975-02-16T01:00:00"
#>  [70] "1975-05-16T01:00:00" "1975-08-16T01:00:00" "1975-11-16T01:00:00"
#>  [73] "1976-02-16T01:00:00" "1976-05-16T01:00:00" "1976-08-16T01:00:00"
#>  [76] "1976-11-16T01:00:00" "1977-02-16T01:00:00" "1977-05-16T02:00:00"
#>  [79] "1977-08-16T02:00:00" "1977-11-16T01:00:00" "1978-02-16T01:00:00"
#>  [82] "1978-05-16T02:00:00" "1978-08-16T02:00:00" "1978-11-16T01:00:00"
#>  [85] "1979-02-16T01:00:00" "1979-05-16T02:00:00" "1979-08-16T02:00:00"
#>  [88] "1979-11-16T01:00:00" "1980-02-16T01:00:00" "1980-05-16T02:00:00"
#>  [91] "1980-08-16T02:00:00" "1980-11-16T01:00:00" "1981-02-16T01:00:00"
#>  [94] "1981-05-16T02:00:00" "1981-08-16T02:00:00" "1981-11-16T01:00:00"
#>  [97] "1982-02-16T01:00:00" "1982-05-16T02:00:00" "1982-08-16T02:00:00"
#> [100] "1982-11-16T01:00:00" "1983-02-16T01:00:00" "1983-05-16T02:00:00"
#> [103] "1983-08-16T02:00:00" "1983-11-16T01:00:00" "1984-02-16T01:00:00"
#> [106] "1984-05-16T02:00:00" "1984-08-16T02:00:00" "1984-11-16T01:00:00"
#> [109] "1985-02-16T01:00:00" "1985-05-16T02:00:00" "1985-08-16T02:00:00"
#> [112] "1985-11-16T01:00:00" "1986-02-16T01:00:00" "1986-05-16T02:00:00"
#> [115] "1986-08-16T02:00:00" "1986-11-16T01:00:00" "1987-02-16T01:00:00"
#> [118] "1987-05-16T02:00:00" "1987-08-16T02:00:00" "1987-11-16T01:00:00"
#> [121] "1988-02-16T01:00:00" "1988-05-16T02:00:00" "1988-08-16T02:00:00"
#> [124] "1988-11-16T01:00:00" "1989-02-16T01:00:00" "1989-05-16T02:00:00"
#> [127] "1989-08-16T02:00:00" "1989-11-16T01:00:00" "1990-02-16T01:00:00"
#> [130] "1990-05-16T02:00:00" "1990-08-16T02:00:00" "1990-11-16T01:00:00"
#> [133] "1991-02-16T01:00:00" "1991-05-16T02:00:00" "1991-08-16T02:00:00"
#> [136] "1991-11-16T01:00:00" "1992-02-16T01:00:00" "1992-05-16T02:00:00"
#> [139] "1992-08-16T02:00:00" "1992-11-16T01:00:00" "1993-02-16T01:00:00"
#> [142] "1993-05-16T02:00:00" "1993-08-16T02:00:00" "1993-11-16T01:00:00"
#> [145] "1994-02-16T01:00:00" "1994-05-16T02:00:00" "1994-08-16T02:00:00"
#> [148] "1994-11-16T01:00:00" "1995-02-16T01:00:00" "1995-05-16T02:00:00"
#> [151] "1995-08-16T02:00:00" "1995-11-16T01:00:00" "1996-02-16T01:00:00"
#> [154] "1996-05-16T02:00:00" "1996-08-16T02:00:00" "1996-11-16T01:00:00"
#> [157] "1997-02-16T01:00:00" "1997-05-16T02:00:00" "1997-08-16T02:00:00"
#> [160] "1997-11-16T01:00:00" "1998-02-16T01:00:00" "1998-05-16T02:00:00"
#> [163] "1998-08-16T02:00:00" "1998-11-16T01:00:00" "1999-02-16T01:00:00"
#> [166] "1999-05-16T02:00:00" "1999-08-16T02:00:00" "1999-11-16T01:00:00"
#> [169] "2000-02-16T01:00:00" "2000-05-16T02:00:00" "2000-08-16T02:00:00"
#> [172] "2000-11-16T01:00:00" "2001-02-16T01:00:00" "2001-05-16T02:00:00"
#> [175] "2001-08-16T02:00:00" "2001-11-16T01:00:00" "2002-02-16T01:00:00"
#> [178] "2002-05-16T02:00:00" "2002-08-16T02:00:00" "2002-11-16T01:00:00"
#> [181] "2003-02-16T01:00:00" "2003-05-16T02:00:00" "2003-08-16T02:00:00"
#> [184] "2003-11-16T01:00:00" "2004-02-16T01:00:00" "2004-05-16T02:00:00"
#> [187] "2004-08-16T02:00:00" "2004-11-16T01:00:00" "2005-02-16T01:00:00"
#> [190] "2005-05-16T02:00:00" "2005-08-16T02:00:00" "2005-11-16T01:00:00"
#> [193] "2006-02-16T01:00:00" "2006-05-16T02:00:00" "2006-08-16T02:00:00"
#> [196] "2006-11-16T01:00:00" "2007-02-16T01:00:00" "2007-05-16T02:00:00"
#> [199] "2007-08-16T02:00:00" "2007-11-16T01:00:00" "2008-02-16T01:00:00"
#> [202] "2008-05-16T02:00:00" "2008-08-16T02:00:00" "2008-11-16T01:00:00"
#> [205] "2009-02-16T01:00:00" "2009-05-16T02:00:00" "2009-08-16T02:00:00"
#> [208] "2009-11-16T01:00:00" "2010-02-16T01:00:00" "2010-05-16T02:00:00"
#> [211] "2010-08-16T02:00:00" "2010-11-16T01:00:00" "2011-02-16T01:00:00"
#> [214] "2011-05-16T02:00:00" "2011-08-16T02:00:00" "2011-11-16T01:00:00"
#> [217] "2012-02-16T01:00:00" "2012-05-16T02:00:00" "2012-08-16T02:00:00"
#> [220] "2012-11-16T01:00:00" "2013-02-16T01:00:00" "2013-05-16T02:00:00"
#> [223] "2013-08-16T02:00:00" "2013-11-16T01:00:00" "2014-02-16T01:00:00"
#> [226] "2014-05-16T02:00:00" "2014-08-16T02:00:00" "2014-11-16T01:00:00"
#> [229] "2015-02-16T01:00:00" "2015-05-16T02:00:00" "2015-08-16T02:00:00"
#> [232] "2015-11-16T01:00:00" "2016-02-16T01:00:00" "2016-05-16T02:00:00"
#> [235] "2016-08-16T02:00:00" "2016-11-16T01:00:00"
#> attr(,"type")
#> temporal_coefficents
#>